+++ components/org.wso2.carbon.identity.application.authenticator.oidc/src/main/java/org/wso2/carbon/identity/application/authenticator/oidc/federatedIdpInitLogout/exception/LogoutServerException.java
@@ -16,31 +16,29 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.application.authenticator.oidc;
-
-import org.wso2.carbon.identity.application.authentication.framework.exception.FrameworkException;
+package org.wso2.carbon.identity.application.authenticator.oidc.federatedIdpInitLogout.exception;
 
 /**
- * Handle logout related errors
+ * This class handle server side errors
  */
-public class LogoutException extends FrameworkException {
+public class LogoutServerException extends LogoutException {
 
-    public LogoutException(String message) {
+    public LogoutServerException(String message) {
 
         super(message);
     }
 
-    public LogoutException(String errorCode, String message) {
+    public LogoutServerException(String errorCode, String message) {
 
         super(errorCode, message);
     }
 
-    public LogoutException(String message, Throwable cause) {
+    public LogoutServerException(String message, Throwable cause) {
 
         super(message, cause);
     }
 
-    public LogoutException(String errorCode, String message, Throwable cause) {
+    public LogoutServerException(String errorCode, String message, Throwable cause) {
 
         super(errorCode, message, cause);
     }
